requirements.txt
fastapi
uvicorn
torch
flwr
pandas
numpy
scikit-learn
pydantic
email-validator
python-jose[cryptography]
